refactor(plasma-ui,plasma-core,plasma-hope): Refactor some components for clean re-export
