+++ packages/plasma-core/src/components/Typography/Button.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import { button1, button2 } from '../../../tokens';
-import { applyHyphens, applySpacing, BreakWordProps, SpacingProps } from '../../../mixins';
+import { button1, button2 } from '../../tokens';
+import { applyHyphens, applySpacing, BreakWordProps, SpacingProps } from '../../mixins';
 
 export const Button1 = styled.div<SpacingProps & BreakWordProps>`
     ${applyHyphens}
