+++ packages/plasma-core/src/components/Typography/Caption.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import { caption } from '../../../tokens';
-import { applyHyphens, applySpacing, BreakWordProps, SpacingProps } from '../../../mixins';
+import { caption } from '../../tokens';
+import { applyHyphens, applySpacing, BreakWordProps, SpacingProps } from '../../mixins';
 
 export const Caption = styled.div<SpacingProps & BreakWordProps>`
     ${applyHyphens}
