+++ packages/plasma-core/src/components/Typography/Display.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import { display1, display2, display3 } from '../../../tokens';
-import { applyHyphens, applySpacing, BreakWordProps, SpacingProps } from '../../../mixins';
+import { display1, display2, display3 } from '../../tokens';
+import { applyHyphens, applySpacing, BreakWordProps, SpacingProps } from '../../mixins';
 
 export const Display1 = styled.div<SpacingProps & BreakWordProps>`
     ${applyHyphens}
