+++ packages/plasma-core/src/components/Typography/Footnote.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import { footnote1, footnote2 } from '../../../tokens';
-import { applyHyphens, applySpacing, BreakWordProps, SpacingProps } from '../../../mixins';
+import { footnote1, footnote2 } from '../../tokens';
+import { applyHyphens, applySpacing, BreakWordProps, SpacingProps } from '../../mixins';
 
 export const Footnote1 = styled.div<SpacingProps & BreakWordProps>`
     ${applyHyphens}
