+++ packages/plasma-core/src/components/Typography/Underline.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import { underline } from '../../../tokens';
-import { applyHyphens, applySpacing, BreakWordProps, SpacingProps } from '../../../mixins';
+import { underline } from '../../tokens';
+import { applyHyphens, applySpacing, BreakWordProps, SpacingProps } from '../../mixins';
 
 export const Underline = styled.div<SpacingProps & BreakWordProps>`
     ${applyHyphens}
