+++ packages/plasma-hope/src/components/Select/index.ts
@@ -2,4 +2,4 @@ export { Select } from './Select';
 export type { SelectProps } from './Select';
 export type { SelectRefElement } from './SelectButton';
 
-export { SelectGroup } from './views/b2c/SelectGroup';
+export { SelectGroup } from './SelectGroup';
