+++ packages/plasma-ui/src/components/Icon/index.ts
@@ -1,0 +1,2 @@
+export { IconSet } from './IconSet';
+export type { IconSetProps } from './IconSet';
